% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_history.R
\name{location_history}
\alias{location_history}
\title{Get location history}
\usage{
location_history(session, date, asKML = FALSE)
}
\arguments{
\item{session}{An rvest session returned from the login() function}

\item{date}{A date string in YYYY-MM-DD format}

\item{asKML}{TRUE if desired output is KML, FALSE if desired output is a dataframe of data parsed from the KML}
}
\value{
A dataframe of location data or raw KML text
}
\description{
After login, get location history for a specified date
}
\examples{
\dontrun{
sess <- login(username="corynissen@gmail.com", password="my_google_password")
df <- location_history(session=sess, date="2016-04-04")
}
location_history
}
\keyword{login}

