% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to Google}
\usage{
login(username, password)
}
\arguments{
\item{username}{Your Google account username or email address}

\item{password}{Your Google account password}
}
\value{
An rvest session
}
\description{
Login to Google
}
\examples{
\dontrun{
sess <- login(username="corynissen@gmail.com", password="my_google_password")
}
login
}
\keyword{login}

